Modify `sort` function signature
If the underlying data source has some kind of sorting functionality built in, it's unlikely a full module will be required. In the case of Ecto, for instance, the resulting sorters are 100% compatible and there's no reason to force the creation of a whole module when a single function will do the trick.