Remove assumption about atom field names
Initially, collection, filter, and sort made assumptions about atoms being the underlying field names and converted strings accordingly. Not only is this an unnecessary limitation, but a lot of automatic conversion of "convenience" values (e.g. `-field_name` for a descending sort) are really about dealing with data from the edge rather than having a generic interface to a collection.