Rename `to_*` to `cast` and add :ok/:error
The "cast" naming is consistent with Ecto and with modules that will be added in the near future. Invalid input should now result in `:error` tuples which can be handled without exceptions. Additionally, `:binary` was renamed to `:string` because on the edge, the datatype will almost always be referred to as a string.