Allow types to modify queries before applying
This was mainly added because no matter what, Ecto will have atoms for keys. Rather than force the input to be atoms, the type can now transform input before it is set to the actual data source. This will generally result in some kind of string to atom or string inflection change.