Add JSON:API number/size pagination strategy
After working on this it's clear that the pagination functionality on collections themselves will have to support multiple strategies. This module is a little overly complicated for this single strategy, but it was originally written in mind with more than one. I just don't have the time or desire to work on it just yet.