Clarify to pagination functions
For APIs, getting pagination info along with the total is a pretty common thing. `all_with_page_info/2` handles this. `filter_and_sort/2` has been made the public function `query/2` which is consistent with the naming in `Resource`. This was also done to allow datasets which are lazy in nature to have multiple queries applied before ultimately using them to return lists.