Rename `Resource` to `Type`
The name `Resource` was a bad call because, in practice, the "resource" is generally the data that's going to be represented by a particular type. This made working with resource in terms of naming kind of a pain. `Resourceful.Type` with a `name` rather than a `resource_type` is much clearer. This required renaming a lot of functions throughout.