Remove `getter` and `map_value` from attributes
Getting values is really done from the perspective of the type and when dealing with nesting, `map_value` complicates things. It was meant to be used in conjunction with a custom `getter` which is a feature that adds a lot of complexity (and is flat out broken with registries) for something that's better handled at the data source layer. Attributes now just describe the query and mapping rules around an attribute.