Add relationships and namespaced fields
Types can now define relationships with other types. This absolutely requires the use of a registry in order to work. Fields can be namespaced in queries, at least for Ecto, with a named join. This also includes some hackery to introspect and dynamically name joins on Ecto queries. Finally, a number of modules got documentation and typespec updates. Private functions are no longer necessarily at the bottom and are instead generally placed in proximity to the public functions that utilize them. This is purely stylistic.