Query types without registries again
Types without registries were broken for queries with the addition of relationships. This has been fixed. In addition, sort and filter delegates are "aware" of attributes and graphed fields now and can coerce them directly into something native to the dataset. In theory, even lists of nested resources can be sorted and filtered. (Gotta write tests.)